# OEUPB-Docs/specs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OE UPB - Observatorio de Egresados",
    "description": "API RESTful para el procesamiento de egresados UPB.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/reportes/general": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Get Reporte General",
        "operationId": "get_reporte_general_api_reportes_general_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KpisResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/reportes/tendencias": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Get Tendencias",
        "operationId": "get_tendencias_api_reportes_tendencias_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "indicador",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "empleabilidad",
                "salario",
                "satisfaccion"
              ],
              "type": "string",
              "default": "empleabilidad",
              "title": "Indicador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TendenciasResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reportes/explorador/init": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Explorador Init",
        "operationId": "explorador_init_api_reportes_explorador_init_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExploradorInitResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/reportes/explorador": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Explorador Data",
        "operationId": "explorador_data_api_reportes_explorador_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pregunta",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 300,
              "title": "Pregunta"
            }
          },
          {
            "name": "momento",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Momento 0, 1 o 5",
              "title": "Momento"
            },
            "description": "Momento 0, 1 o 5"
          },
          {
            "name": "programa",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 150
                },
                {
                  "type": "null"
                }
              ],
              "title": "Programa"
            }
          },
          {
            "name": "anio",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2200,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "Anio"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExploradorResponse"
                }
              }
            }
          },
          "422": {
            "description": "La variable no pertenece al catálogo analítico autorizado o los filtros son inválidos"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChartDatasetResponse": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "data": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "Data"
          },
          "borderColor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bordercolor"
          },
          "backgroundColor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backgroundcolor"
          },
          "borderWidth": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Borderwidth"
          },
          "pointBackgroundColor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pointbackgroundcolor"
          },
          "pointBorderColor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pointbordercolor"
          },
          "pointBorderWidth": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pointborderwidth"
          },
          "pointRadius": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pointradius"
          },
          "pointHoverRadius": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pointhoverradius"
          },
          "fill": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fill"
          },
          "tension": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tension"
          },
          "spanGaps": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spangaps"
          }
        },
        "type": "object",
        "required": [
          "label",
          "data"
        ],
        "title": "ChartDatasetResponse"
      },
...
      "ExploradorInitResponse": {
        "properties": {
          "preguntas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preguntas"
          },
          "programas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Programas"
          },
          "anios": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Anios"
          }
        },
        "type": "object",
        "required": [
          "preguntas",
          "programas",
          "anios"
        ],
        "title": "ExploradorInitResponse"
      },
      "ExploradorResponse": {
        "properties": {
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Labels"
          },
          "valores": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Valores"
          }
        },
        "type": "object",
        "required": [
          "labels",
          "valores"
        ],
        "title": "ExploradorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KpisResponse": {
        "properties": {
          "total_egresados": {
            "type": "integer",
            "title": "Total Egresados"
          },
          "tasa_empleabilidad": {
            "type": "number",
            "title": "Tasa Empleabilidad"
          },
          "promedio_salarial": {
            "type": "number",
            "title": "Promedio Salarial"
          },
          "distribucion_programas": {
            "additionalProperties": true,
            "type": "object",
            "title": "Distribucion Programas"
          },
          "nivel_satisfaccion": {
            "additionalProperties": true,
            "type": "object",
            "title": "Nivel Satisfaccion"
          }
        },
        "type": "object",
        "required": [
          "total_egresados",
          "tasa_empleabilidad",
          "promedio_salarial",
          "distribucion_programas",
          "nivel_satisfaccion"
        ],
        "title": "KpisResponse"
      },
...
      "TendenciasResponse": {
        "properties": {
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Labels"
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/ChartDatasetResponse"
            },
            "type": "array",
            "title": "Datasets"
          }
        },
        "type": "object",
        "required": [
          "labels",
          "datasets"
        ],
        "title": "TendenciasResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}